+++ forms-flow-admin/src/declarations.d.ts
@@ -39,5 +39,5 @@ declare module "*.svg" {
 }
 
 declare module "@formsflow/service" {
-	export const {KeycloakService, RequestService, StorageService}: any;
+	export const {KeycloakService, RequestService, StorageService,i18nService}: any;
 }