+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog for formsflow.ai
 
-Mark items as `Added`, `Changed`, `Fixed`, `Removed`, `Untested Features`, `Upcoming Features`, `Known Issues`
+Mark items as `Added`, `Changed`, `Fixed`, `Modified`, `Removed`, `Untested Features`, `Upcoming Features`, `Known Issues`
+
+## 5.3.0 - 2023-11-24
+
+`Added`
+
+**forms-flow-service**
+  - Added date and time service
+  - Added resource bundle 
+
+`Modified`
+
+**forms-flow-admin**
+ - Modified user interface and improved user experience 
+ - Modified footer style
+ 
+**forms-flow-nav**
+ - Modified navbar style
+   
+**forms-flow-theme**
+ - Modified CSS to SCSS and folder structure
+ - Modified toast color override
+ 
 
 ## 5.2.0 - 2022-07-07